Moved Pin Definitions to Header
Moved the pin define statements to a header file. I couldn't find a hardware version number, so I just called it "Teensy-36-Mounted" because the Teensy 3.6 is mounted directly to the board, not embedded into it. I'm not married to the name, just the first thing I came up with